backend/foodgram/requirements.txt
requests==2.26.0
asgiref==3.2.10
Django==2.2.19
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pillow==9.1.1
sorl-thumbnail==12.8.0
drf_extra_fields==3.4.0